changed long double back to double now using distance to lower band radius as secondary lower bound

+++ networkit/cpp/Globals.h
@@ -36,9 +36,9 @@ namespace NetworKit {
 }
 
 #ifdef __INTEL_COMPILER
-constexpr long double PI = 3.141592653589793238462643383279502884197169399375105820974944592307816406286;
+constexpr double PI = 3.141592653589793238462643383279502884197169399375105820974944592307816406286;
 #else
-const long double PI = 2.0*std::acos(0);
+const double PI = 2.0*std::acos(0);
 #endif
 
 // CODE STYLE GUIDELINES: Do not rely on global variables for algorithm parametrization.
